Tweaks to support serialization into puya

+++ src/index.ts
@@ -10,6 +10,7 @@ import { registerPTypes } from './awst_build/ptypes/register'
 import { typeRegistry } from './awst_build/type-registry'
 import { generateTempFile } from './util/generate-temp-file'
 import { jsonSerializeAwst } from './awst/json-serialize-awst'
+import { jsonSerializeSourceFiles } from './parser/json-serialize-source-files'
 
 export type CompileResult = {
   logs: LogEvent[]
@@ -37,10 +38,14 @@ export function compile(options: CompileOptions): CompileResult {
     }
     throw e
   }
-
-  const moduleAwstFile = generateTempFile()
-  logger.info(undefined, `Writing awst to ${moduleAwstFile.filePath}`)
-  moduleAwstFile.writeFileSync(jsonSerializeAwst(moduleAwst), 'utf-8')
+  if (!options.dryRun) {
+    const moduleAwstFile = generateTempFile()
+    logger.info(undefined, `Writing awst to ${moduleAwstFile.filePath}`)
+    moduleAwstFile.writeFileSync(jsonSerializeAwst(moduleAwst), 'utf-8')
+    const moduleSourceFile = generateTempFile()
+    logger.info(undefined, `Write source to ${moduleSourceFile.filePath}`)
+    moduleSourceFile.writeFileSync(jsonSerializeSourceFiles(programResult.sourceFiles), 'utf-8')
+  }
 
   return {
     programDirectory,
